Setup websocket connection between ctl and server.
Currently this just ping/pongs between the ctl and server but is ready to add on proxy and api.
Also added a pool for the connections which is quite basic atm but will be used to check if a connection exists for a subdomain when a user hits the server.