Can now send http requests to backend
Have updated the tunnel_handler to read in the request and send it through the respective websocket connection to the client.
Although this is nice may consider pivoting to just a pure TCP connection rather than gin and websocekts as this is a bit limiting in functionality. Will need to do some more digging