Started migration to tcp connections
Have changed course again and decided to make the TCP connection from the ctl command to the server instead of the other way around. This is to get around NAT primarily which would have been a challenge with running the server on the ctl host.
This is still a draft and can see connectivity but currently dealing with an issue where the first request succeeds but subsequent requests fail.