Can initailize a tcp connection e2e
Whilst this is good, the connection gets dropped after sending through the password for the db... Will need to continue to work on this and improve the overall architecture of the app. The different tcp server files is confusing, and the StreamingConnection interface is a bit rough. Plus ill need to figure out how to get the hostname when making tcp connections!!